Refactor registry classes for fuels and fluids
