+++ src/main/java/com/jusipat/castleblocks/CastleBlocksMod.java
@@ -1,9 +1,9 @@
 package com.jusipat.castleblocks;
 
-import com.jusipat.castleblocks.block.fluid.ModFluids;
+import com.jusipat.castleblocks.registry.ModFluids;
 import com.jusipat.castleblocks.registry.ModBlocks;
 import com.jusipat.castleblocks.registry.ModItems;
-import com.jusipat.castleblocks.registry.ModRegistries;
+import com.jusipat.castleblocks.registry.ModFuels;
 import net.fabricmc.api.ModInitializer;
 import net.fabricmc.fabric.api.blockrenderlayer.v1.BlockRenderLayerMap;
 import net.fabricmc.fabric.api.client.itemgroup.FabricItemGroupBuilder;
@@ -25,7 +25,7 @@ public class CastleBlocksMod implements ModInitializer {
 	public void onInitialize() {
 		ModBlocks.registerBlocks();
 		ModItems.registerItems();
-		ModRegistries.registerModFuels();
+		ModFuels.registerFuels();
 
 		BlockRenderLayerMap.INSTANCE.putBlock(ModBlocks.CASTLE_DOOR, RenderLayer.getCutout());
 		FluidRenderHandlerRegistry.INSTANCE.register(ModFluids.PITCH_STILL,
