--- src/main/java/com/jusipat/castleblocks/block/fluid/ModFluidBlock.java
@@ -1,10 +1,0 @@
-package com.jusipat.castleblocks.block.fluid;
-
-import net.minecraft.block.FluidBlock;
-import net.minecraft.fluid.FlowableFluid;
-
-public class ModFluidBlock extends FluidBlock {
-	public ModFluidBlock(FlowableFluid fluid, Settings settings) {
-		super(fluid, settings);
-	}
-}
