+++ src/main/java/com/jusipat/castleblocks/registry/ModFuels.java
@@ -2,8 +2,8 @@ package com.jusipat.castleblocks.registry;
 
 import net.fabricmc.fabric.api.registry.FuelRegistry;
 
-public class ModRegistries {
-	public static void registerModFuels() {
+public class ModFuels {
+	public static void registerFuels() {
 		FuelRegistry registry = FuelRegistry.INSTANCE;
 
 		registry.add(ModItems.PITCH, 2200);
